Adapt AsiceArchiveTest to test no signature when certificate holder is null
There was already clearly support in the library for suppressing the creation
of a signature file when the AsiceBuilder.Create method is called with a null
value for its ICertificateHolder signCertificate parameter. This didn't seem
to be tested. This change adds a new test that calls the existing
AsiceArchiveTest.TestArchive(ICertificateHolder certHolder) helper method with
the parameter set to null. This helper method is modified so it replicates the
ternary expression found in AsiceBuilder that switches between

 * CadesManifestCreator.CreateWithoutSignatureFile()
 * CadesManifestCreator.CreateWithSignatureFile()

so that CadesManifestCreator is created to reflect realistic behaviour
depending on whether certHolder is null.

Some expectancies in the helper method are updated, so that the total number
of files in the container is lower if there is no signature, and so that no
signature is looked for when it shouldn't be there.