Clean up ISignatureCreator to avoid mentioning CAdES, consolidate to one method
In preparation for introducing XAdES support, the interface ISignatureCreator
and the implementing class SignatureCreator are changed from having two methods
-- CreateSignatureFile() taking an enumeration of the container entries but
throwing NotImplementedException in the implementation and the other being
named CreateCadesSignatureFile() which mentions CAdES by name and which takes
ManifestContainer as parameter -- to just having a single CreateSignatureFile()
which takes both of these parameters.

The implementation code for CreateCadesSignatureFile() is retained, unchanged
but now in the method named CreateSignatureFile(). The enumeration of entries
is unused in the current CAdES-specific implementation, but will become
important in the forthcoming XAdES implementation of ISignatureCreator.