In AsiceArchive, move logic for adding signature into new AddSignature method
In preparation for introducing XAdES support, the logic for generating a file
containing the signature and adding it as an entry in the container is now
moved from AddManifest(), where it sat in a conditional guarded by a clause
which included the term 'manifest.ManifestSpec == ManifestSpec.Cades', into a
new method AddSignature().

The manifest's own AsicePackageEntry is now made available in a new property
on ManifestContainer called PackageEntry. This is set in AddManifest().

In the Dispose() function in AsiceArchive, where AddManifest() was already
called, the new AddSignature() method is also called. AddManifest() goes from
being void to returning the ManifestContainer it creates, which is passed to
AddSignature() so that the manifest can be signed.

A new string constant is added to AsiceConstants for "application/xml". This
is a string literal rather than referencing MediaTypeNames.Application.Xml,
since Visual Studio marks use of that symbol as an error as long as
TargetFrameworks includes "netstandard2.0".