Add support for certificate chains in ICertificateHolder, needed for XAdES
In preparation for the introduction of XAdES support, a new method
GetCertificateChain() is added to ICertificateHolder, and the main
implementation PreloadedCertificateHolder gets a new optional parameter in one
of its two exposed Create() methods for passing in the certificate chain.

The other implementation, SystemX509CertificateHolder, implements the new
interface method simply by returning an empty array.

Certificate chain support is needed since some recipients of ASiC-E containers
signed with XAdES require a full chain of certificates to be included in the
signature file, up to and including a trusted root certificate.