Add the XAdES signature implementation to library
The implementation of XAdES signatures is located in the new class
XadesSignatureCreator.

The implementation makes use the XML canonicalization (C14N) algorithm located
in System.Security.Cryptography.Xml.XmlDsigC14NTransform, so therefore
System.Security.Cryptography.Xml has been added as a dependency in the project.

Other changes in this commit:
 * In AsicePackageEntry, it introduces an ID property, backed by a static
   identCounter which is incremented in the constructor, since XAdES requires
   a unique ID for each entry in the signature file.
 * In AsiceConstants, a new static readonly MimeType for the XAdES signature
   is introduced, referencing the const string ContentTypeApplicationXml
   introduced previously.
 * A new XadesSignature class is introduced, implementing the same interface,
   ISignatureFileRefCreator, as the CadesSignature class.