In AsiceBuilder, add Xades in two of three Find..() helper methods
In FindSignatureFileRefCreator() and FindSignatureCreator(), the enum value
for XAdES is added to the pattern-matching expressions.

This is not done for FindManifestCreator() however, since a separate manifest
format for XAdES is not implemented here. The initial usecase for XAdES is to
support the Norwegian governmental DPI system, which uses a completely custom
manifest format. Support for custom manifest formats comes in the next commit.