Add optional makeManifestCreator Func param to AsiceBuilder.Create
To faciliate the need to make a custom manifest format, which was critical in
the first usecase for XAdES support, this change adds an optional parameter to
AsiceBuilder.Create(), which is default null. A Func<IManifestCreator> object
can be passed here, and it will be used in place of the library's default
manifest according to the selected ManifestSpec.

IManifestCreator is already a public interface, but with this change it is
expected that some integrators will need to implement it themselves. API
documentation has therefore been added in the interface.