Cleanup: Remove ICertificateHolder knowledge from AsiceArchive
The _signatureCertificate member in AsiceArchive was now unused, and this
change removes it, along with the ICertificateHolder parameters in each of the
constructors. The constructors were then further simplified by removing the
verification that the certificate holder needed to be null if and only if the
signature file ref creator was null.