Add tests for custom manifest formats in AsiceBuilder.Create()
A new test file is added that uses AsiceBuilder.Create() with the new optional
makeManifestCreator parameter of type Func<IManifestCreator>. The test includes
a class that implements IManifestCreator, and illustrates a bespoke manifest
format.

It asserts that it finds the manifest file at a custom location, containing the
exact XML that it should produce. It also asserts that a custom manifest does
not affect the location of the signature file.