Dependency-inject UtcNow into XadesSignatureCreator using Func<DateTime> param
To facilitate testability, instead of having the new class for creating XAdES
signatures, XadesSignatureCreator, use DateTime.UtcNow directly, it will now
call a member field of type Func<DateTime> called _getUtcNow. This will default
to DateTime.UtcNow, but can also be passed in via AsiceBuilder.Create(), which
sends the value through the static Create method and constructor of
XadesSignatureCreator.

Unlike CAdES, the XAdES logic needs to access the current time in order to
include this in the signature file itself.

In tests, the parameter can be a Func that simply returns the same DateTime
value each time. With this change, such a value is passed in via TestArchive()
in AsiceArchiveTest.